requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.3.0
django-tables2==2.3.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.1
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.2.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==6.2.1
postgres==3.0.0
progress==1.5
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2019.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
zipp==3.1.0
